`/deck/id` page: add CardPreview
